refactor(agent): move local agent runtime
