+++ examples/agentic/coding/agent_loop.py
@@ -1,3 +1,3 @@
 """Compatibility imports for the moved coding-agent loop."""
 
-from areno.agentic.coding.agent_loop import *  # noqa: F403
+from areno.agent.agent_loop import *  # noqa: F403
