+++ examples/agentic/coding/coding_tools.py
@@ -1,3 +1,3 @@
 """Compatibility imports for the moved coding-agent tools."""
 
-from areno.agentic.coding.coding_tools import *  # noqa: F403
+from areno.agent.tools import *  # noqa: F403
